[test] Revert to 09af6144492f1b5ee7164a1227a5669817e986de

+++ cpp_src/client/rpcclient.h
@@ -104,7 +104,7 @@ protected:
 	Namespace *getNamespace(string_view nsName);
 	Error startWorkers();
 	Error addConnectEntry(const string &dsn, const client::ConnectOpts &opts, size_t idx);
-	void run(int thIdx);
+	void run(size_t thIdx);
 	void onUpdates(net::cproto::RPCAnswer &ans, cproto::ClientConnection *conn);
 	bool onConnectionFail(int failedDsnIndex);
